Extract to independent "block" per @phamann
